+++ web/vote-pro-pc/package.json
@@ -4,43 +4,24 @@
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "build": "vue-cli-service build"
   },
   "dependencies": {
     "axios": "^0.19.2",
-    "bcx-api": "^2.1.21",
     "core-js": "^3.6.4",
     "element-ui": "^2.13.0",
     "vue": "^2.6.11",
-    "vue-axios": "^2.1.5",
-    "vue-i18n": "^8.15.4",
-    "vue-router": "^3.1.5",
-    "vue-smooth-scroll": "^1.0.13"
+    "vue-router": "^3.1.5"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.2.0",
-    "@vue/cli-plugin-eslint": "~4.2.0",
     "@vue/cli-plugin-router": "~4.2.0",
     "@vue/cli-service": "~4.2.0",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.1.2",
-    "vue-template-compiler": "^2.6.11"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
+    "vue-template-compiler": "^2.6.11",
+    "bcx-api": "^2.0.2",
+    "vue-axios": "^2.1.5",
+    "vue-i18n": "^8.12.0",
+    "axios": "^0.19.0"
   },
   "browserslist": [
     "> 1%",
