--- web/vote-pro-wap/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <!-- <meta name="viewport" content="width=device-width,initial-scale=1.0"> -->
-    <meta http-equiv="Pragma" content="no-cache">
-    <meta http-equiv="Cache-Control" content="no-cache">
-    <meta http-equiv="Expires" content="0">
-    <link rel="icon" href="https://jdi.cocosbcx.net/image/crypto_knights_web/crypto_knights.png">
-    <meta name="viewport" content="width=device-width,maximum-scale=1.0,minimum-scale=1.0,initial-scale=1.0,user-scalable=no">
-    <title>Vote</title>
-  </head>
-  <body>
-    
-    <div id="app"></div>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
