+++ web/vote-pro-wap/src/router/index.js
@@ -1,52 +1,57 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from '@/components/Home'
-import Test from '@/components/Test'
-import LookForwardTo from '@/components/LookForwardTo'
-import QuestionnaireIntro from '@/components/QuestionnaireIntro'
-import QuestionnaireForm from '@/components/QuestionnaireForm'
-import QuestionnaireConfirm from '@/components/QuestionnaireConfirm'
-import QuestionnaireSuccess from '@/components/QuestionnaireSuccess'
+import Home from '@/views/Home'
+import Test from '@/views/Test'
+import LookForwardTo from '@/views/LookForwardTo'
+import QuestionnaireIntro from '@/views/QuestionnaireIntro'
+import QuestionnaireForm from '@/views/QuestionnaireForm'
+import QuestionnaireConfirm from '@/views/QuestionnaireConfirm'
+import QuestionnaireSuccess from '@/views/QuestionnaireSuccess'
 
 Vue.use(Router)
 
-export default new Router({
+const routes = [
+  {
+    path: '/LookForwardTo',
+    name: 'LookForwardTo',
+    component: LookForwardTo
+  },
+  {
+    path: '/',
+    name: 'Home',
+    component: Home
+  },
+  {
+    path: '/test',
+    name: 'Test',
+    component: Test
+  },
+  {
+    path: '/questionnaireintro',
+    name: 'QuestionnaireIntro',
+    component: QuestionnaireIntro
+  },
+  {
+    path: '/questionnaireform',
+    name: 'QuestionnaireForm',
+    component: QuestionnaireForm
+  },
+  {
+    path: '/auestionnaireconfirm',
+    name: 'QuestionnaireConfirm',
+    component: QuestionnaireConfirm
+  },
+  {
+    path: '/questionnairesuccess',
+    name: 'QuestionnaireSuccess',
+    component: QuestionnaireSuccess
+  }
+]
+
+const router = new Router({
   mode: 'history',
-  routes: [
-    {
-      path: '/LookForwardTo',
-      name: 'LookForwardTo',
-      component: LookForwardTo
-    },
-    {
-      path: '/',
-      name: 'Home',
-      component: Home
-    },
-    {
-      path: '/test',
-      name: 'Test',
-      component: Test
-    },
-    {
-      path: '/questionnaireintro',
-      name: 'QuestionnaireIntro',
-      component: QuestionnaireIntro
-    },
-    {
-      path: '/questionnaireform',
-      name: 'QuestionnaireForm',
-      component: QuestionnaireForm
-    },
-    {
-      path: '/auestionnaireconfirm',
-      name: 'QuestionnaireConfirm',
-      component: QuestionnaireConfirm
-    },
-    {
-      path: '/questionnairesuccess',
-      name: 'QuestionnaireSuccess',
-      component: QuestionnaireSuccess
-    }
-  ]
+  base: process.env.BASE_URL,
+  routes
 })
+
+export default router
